docs(cgroup): documented cgroup source
