+++ sources/cgroup/src/counters.rs
@@ -25,24 +25,30 @@ impl MinMax {
     }
 }
 
+/// Tracks the beginning and latest values of a cumulative counter for a phase.
 #[derive(Debug, Default, Clone, Copy)]
 pub struct BeginEnd(u64, u64);
 
 impl BeginEnd {
+    /// Updates the latest counter value.
     pub fn update(&mut self, value: u64) {
         self.1 = value;
     }
 
+    /// Returns the difference between end and begin values.
     pub fn diff(&self) -> u64 {
         self.1.saturating_sub(self.0)
     }
 
+    /// Starts a new measurement phase by setting begin to end.
     pub fn new_phase(&mut self) {
         self.0 = self.1;
     }
 }
 
+/// Internal trait for phase-based counters.
 trait NewPhase {
+    /// Starts a new measurement phase.
     fn new_phase(&mut self);
 }
 
@@ -60,7 +66,9 @@ impl NewPhase for Option<BeginEnd> {
     }
 }
 
+/// Internal trait for updating optional counters.
 trait UpdateOpt {
+    /// Updates the counter if a value is present.
     fn update(&mut self, value: Option<u64>);
 }
 
@@ -82,18 +90,36 @@ impl UpdateOpt for Option<MinMax> {
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct MemoryCounters {
+    /// Current memory usage.
     pub current: Option<MinMax>,
+
+    /// Current swap usage.
     pub swap_current: Option<MinMax>,
+
+    /// Anonymous memory usage (stack, heap, anon memory mappings).
     pub anon: Option<MinMax>,
+
+    /// File-backed memory usage.
     pub file: Option<MinMax>,
+
+    /// Peak memory usage.
     pub peak: Option<MinMax>,
+
+    /// Shared memory usage.
     pub shmem: Option<MinMax>,
+
+    /// Kernel memory usage.
     pub kernel: Option<MinMax>,
+
+    /// Kernel stack usage.
     pub kernel_stack: Option<MinMax>,
+
+    /// Slab allocator usage.
     pub slab: Option<MinMax>,
 }
 
 impl MemoryCounters {
+    /// Updates counters from a memory snapshot.
     pub fn update(&mut self, snapshot: &MemorySnapshot) {
         self.current.update(snapshot.current);
         self.swap_current.update(snapshot.swap_current);
@@ -106,6 +132,7 @@ impl MemoryCounters {
         self.slab.update(snapshot.slab);
     }
 
+    /// Resets all tracked memory metrics.
     pub fn reset(&mut self) {
         self.current = None;
         self.swap_current = None;
@@ -121,18 +148,33 @@ impl MemoryCounters {
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct CpuCounters {
+    /// Total CPU usage time.
     pub usage_usec: BeginEnd,
+
+    /// User-space CPU usage time.
     pub user_usec: BeginEnd,
+
+    /// Kernel-space CPU usage time.
     pub system_usec: BeginEnd,
 
+    /// Number of scheduler periods.
     pub nr_periods: Option<BeginEnd>,
+
+    /// Number of throttled periods.
     pub nr_throttled: Option<BeginEnd>,
+
+    /// Total throttled time.
     pub throttled_usec: Option<BeginEnd>,
+
+    /// Number of CPU bursts.
     pub nr_bursts: Option<BeginEnd>,
+
+    /// Total burst time.
     pub burst_usec: Option<BeginEnd>,
 }
 
 impl CpuCounters {
+    /// Updates counters from a CPU snapshot.
     pub fn update(&mut self, snapshot: &CpuSnapshot) {
         self.usage_usec.update(snapshot.usage_usec);
         self.user_usec.update(snapshot.user_usec);
@@ -144,6 +186,7 @@ impl CpuCounters {
         self.burst_usec.update(snapshot.burst_usec);
     }
 
+    /// Starts a new measurement phase.
     pub fn new_phase(&mut self) {
         self.usage_usec.new_phase();
         self.user_usec.new_phase();
@@ -158,11 +201,15 @@ impl CpuCounters {
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct IoCounters {
+    /// Bytes read.
     pub rbytes: Option<BeginEnd>,
+
+    /// Bytes written.
     pub wbytes: Option<BeginEnd>,
 }
 
 impl IoCounters {
+    /// Updates counters from an I/O snapshot.
     pub fn update(&mut self, snapshot: &IoSnapshot) {
         if let Some(v) = snapshot.rbytes {
             self.rbytes.get_or_insert_default().update(v);
@@ -173,6 +220,7 @@ impl IoCounters {
         }
     }
 
+    /// Starts a new measurement phase.
     pub fn new_phase(&mut self) {
         if let Some(v) = &mut self.rbytes {
             v.new_phase();
@@ -184,13 +232,20 @@ impl IoCounters {
     }
 }
 
+/// Global application counters.
+///
+/// Stores both process-level and global system metrics.
 #[derive(Debug, Default, Clone, Copy)]
 pub struct Counters {
     pub proc_memory: MemoryCounters,
+
     pub proc_cpu: CpuCounters,
+
     pub proc_io: IoCounters,
 
     pub global_memory: MemoryCounters,
+
     pub global_cpu: CpuCounters,
+
     pub global_io: IoCounters,
 }
